requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-jsonview==2.0.0
Pillow==8.0.1
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
